feat(sidebar): separate DesktopLocalTabContent component for reducing bundle size in web
