+++ surfsense_web/components/layout/ui/sidebar/DocumentsSidebar.tsx
@@ -6,19 +6,17 @@ import {
 	ChevronLeft,
 	ChevronRight,
 	FileText,
-	Folder,
-	FolderPlus,
 	FolderClock,
 	Laptop,
 	Lock,
 	Paperclip,
-	Search,
 	Server,
 	Trash2,
 	Unplug,
 	Upload,
 	X,
 } from "lucide-react";
+import dynamic from "next/dynamic";
 import Link from "next/link";
 import { useParams } from "next/navigation";
 import { useTranslations } from "next-intl";
@@ -49,7 +47,6 @@ import { EXPORT_FILE_EXTENSIONS } from "@/components/shared/ExportMenuItems";
 import {
 	DEFAULT_EXCLUDE_PATTERNS,
 	FolderWatchDialog,
-	type SelectedFolder,
 } from "@/components/sources/FolderWatchDialog";
 import {
 	AlertDialog,
@@ -63,18 +60,8 @@ import {
 } from "@/components/ui/alert-dialog";
 import { Avatar, AvatarFallback, AvatarGroup } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
-import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuLabel,
-	DropdownMenuSeparator,
-	DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import { Drawer, DrawerContent, DrawerHandle, DrawerTitle } from "@/components/ui/drawer";
-import { Input } from "@/components/ui/input";
 import { Skeleton } from "@/components/ui/skeleton";
-import { Separator } from "@/components/ui/separator";
 import { Spinner } from "@/components/ui/spinner";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
@@ -84,7 +71,7 @@ import { getConnectorIcon } from "@/contracts/enums/connectorIcons";
 import type { DocumentTypeEnum } from "@/contracts/types/document.types";
 import { useDebouncedValue } from "@/hooks/use-debounced-value";
 import { useMediaQuery } from "@/hooks/use-media-query";
-import { useElectronAPI } from "@/hooks/use-platform";
+import { usePlatform, useElectronAPI } from "@/hooks/use-platform";
 import { anonymousChatApiService } from "@/lib/apis/anonymous-chat-api.service";
 import { documentsApiService } from "@/lib/apis/documents-api.service";
 import { foldersApiService } from "@/lib/apis/folders-api.service";
@@ -93,8 +80,12 @@ import { authenticatedFetch } from "@/lib/auth-utils";
 import { uploadFolderScan } from "@/lib/folder-sync-upload";
 import { getSupportedExtensionsSet } from "@/lib/supported-extensions";
 import { queries } from "@/zero/queries/index";
-import { LocalFilesystemBrowser } from "./LocalFilesystemBrowser";
 import { SidebarSlideOutPanel } from "./SidebarSlideOutPanel";
+
+const DesktopLocalTabContent = dynamic(
+	() => import("./DesktopLocalTabContent").then((mod) => mod.DesktopLocalTabContent),
+	{ ssr: false }
+);
 
 const NON_DELETABLE_DOCUMENT_TYPES: readonly string[] = ["SURFSENSE_DOCS"];
 const LOCAL_FILESYSTEM_TRUST_KEY = "surfsense.local-filesystem-trust.v1";
@@ -142,9 +133,6 @@ interface WatchedFolderEntry {
 	active: boolean;
 }
 
-const getFolderDisplayName = (rootPath: string): string =>
-	rootPath.split(/[\\/]/).at(-1) || rootPath;
-
 const SHOWCASE_CONNECTORS = [
 	{ type: "GOOGLE_DRIVE_CONNECTOR", label: "Google Drive" },
 	{ type: "GOOGLE_GMAIL_CONNECTOR", label: "Gmail" },
@@ -170,25 +158,40 @@ interface DocumentsSidebarProps {
 
 export function DocumentsSidebar(props: DocumentsSidebarProps) {
 	const isAnonymous = useIsAnonymous();
+	const { isDesktop } = usePlatform();
 	if (isAnonymous) {
 		return <AnonymousDocumentsSidebar {...props} />;
 	}
-	return <AuthenticatedDocumentsSidebar {...props} />;
+	return isDesktop ? (
+		<AuthenticatedDesktopDocumentsSidebar {...props} />
+	) : (
+		<AuthenticatedWebDocumentsSidebar {...props} />
+	);
 }
 
-function AuthenticatedDocumentsSidebar({
+function AuthenticatedDesktopDocumentsSidebar(props: DocumentsSidebarProps) {
+	return <AuthenticatedDocumentsSidebarBase {...props} desktopFeaturesEnabled />;
+}
+
+function AuthenticatedWebDocumentsSidebar(props: DocumentsSidebarProps) {
+	return <AuthenticatedDocumentsSidebarBase {...props} desktopFeaturesEnabled={false} />;
+}
+
+function AuthenticatedDocumentsSidebarBase({
 	open,
 	onOpenChange,
 	isDocked = false,
 	onDockedChange,
 	embedded = false,
 	headerAction,
-}: DocumentsSidebarProps) {
+	desktopFeaturesEnabled,
+}: DocumentsSidebarProps & { desktopFeaturesEnabled: boolean }) {
 	const t = useTranslations("documents");
 	const tSidebar = useTranslations("sidebar");
 	const params = useParams();
 	const isMobile = !useMediaQuery("(min-width: 640px)");
-	const electronAPI = useElectronAPI();
+	const platformElectronAPI = useElectronAPI();
+	const electronAPI = desktopFeaturesEnabled ? platformElectronAPI : null;
 	const searchSpaceId = Number(params.search_space_id);
 	const setConnectorDialogOpen = useSetAtom(connectorDialogOpenAtom);
 	const setRightPanelCollapsed = useSetAtom(rightPanelCollapsedAtom);
@@ -198,9 +201,6 @@ function AuthenticatedDocumentsSidebar({
 
 	const [search, setSearch] = useState("");
 	const debouncedSearch = useDebouncedValue(search, 250);
-	const [localSearch, setLocalSearch] = useState("");
-	const debouncedLocalSearch = useDebouncedValue(localSearch, 250);
-	const localSearchInputRef = useRef<HTMLInputElement>(null);
 	const [activeTypes, setActiveTypes] = useState<DocumentTypeEnum[]>([]);
 	const [filesystemSettings, setFilesystemSettings] = useState<FilesystemSettings | null>(null);
 	const [localTrustDialogOpen, setLocalTrustDialogOpen] = useState(false);
@@ -208,7 +208,7 @@ function AuthenticatedDocumentsSidebar({
 	const [watchedFolderIds, setWatchedFolderIds] = useState<Set<number>>(new Set());
 	const [folderWatchOpen, setFolderWatchOpen] = useAtom(folderWatchDialogOpenAtom);
 	const [watchInitialFolder, setWatchInitialFolder] = useAtom(folderWatchInitialFolderAtom);
-	const isElectron = typeof window !== "undefined" && !!window.electronAPI;
+	const isElectron = desktopFeaturesEnabled && typeof window !== "undefined" && !!window.electronAPI;
 
 	useEffect(() => {
 		if (!electronAPI?.getAgentFilesystemSettings) return;
@@ -1180,161 +1180,24 @@ function AuthenticatedDocumentsSidebar({
 	);
 
 	const localContent = (
-		<div className="flex min-h-0 flex-1 flex-col select-none">
-			<div className="mx-4 mt-4 mb-3">
-				<div className="flex h-7 w-full items-stretch rounded-lg border bg-muted/50 text-[11px] text-muted-foreground">
-					{localRootPaths.length > 0 ? (
-						<DropdownMenu>
-							<DropdownMenuTrigger asChild>
-								<button
-									type="button"
-									className="min-w-0 flex-1 flex items-center gap-1 rounded-l-lg px-2 text-left transition-colors hover:bg-muted/80 focus-visible:outline-none focus-visible:ring-0 focus-visible:ring-offset-0"
-									title={localRootPaths.join("\n")}
-									aria-label="Manage selected folders"
-								>
-									<Folder className="size-3 shrink-0 text-muted-foreground" />
-									<span className="truncate">
-										{localRootPaths.length === 1
-											? "1 folder selected"
-											: `${localRootPaths.length} folders selected`}
-									</span>
-								</button>
-							</DropdownMenuTrigger>
-							<DropdownMenuContent align="start" className="w-56 select-none p-0.5">
-								<DropdownMenuLabel className="px-1.5 pt-1.5 pb-0.5 text-xs font-medium text-muted-foreground">
-									Selected folders
-								</DropdownMenuLabel>
-								<DropdownMenuSeparator className="mx-1 my-0.5" />
-								{localRootPaths.map((rootPath) => (
-									<DropdownMenuItem
-										key={rootPath}
-										onSelect={(event) => event.preventDefault()}
-										className="group h-8 gap-1.5 px-1.5 text-sm text-foreground"
-									>
-										<Folder className="size-3.5 text-muted-foreground" />
-										<span className="min-w-0 flex-1 truncate">
-											{getFolderDisplayName(rootPath)}
-										</span>
-										<button
-											type="button"
-											className="inline-flex size-5 items-center justify-center rounded text-muted-foreground transition-colors hover:text-foreground"
-											onClick={(event) => {
-												event.stopPropagation();
-												void handleRemoveFilesystemRoot(rootPath);
-											}}
-											aria-label={`Remove ${getFolderDisplayName(rootPath)}`}
-										>
-											<X className="size-3" />
-										</button>
-									</DropdownMenuItem>
-								))}
-								<DropdownMenuSeparator className="mx-1 my-0.5" />
-								<DropdownMenuItem
-									variant="destructive"
-									className="h-8 px-1.5 text-xs text-destructive focus:text-destructive"
-									onClick={() => {
-										void handleClearFilesystemRoots();
-									}}
-								>
-									Clear all folders
-								</DropdownMenuItem>
-							</DropdownMenuContent>
-						</DropdownMenu>
-					) : (
-						<div
-							className="min-w-0 flex-1 flex items-center gap-1 px-2"
-							title="No local folders selected"
-						>
-							<Folder className="size-3 shrink-0 text-muted-foreground" />
-							<span className="truncate">No local folders selected</span>
-						</div>
-					)}
-					<Separator
-						orientation="vertical"
-						className="data-[orientation=vertical]:h-3 self-center bg-border"
-					/>
-					{electronAPI ? (
-						<Tooltip>
-							<TooltipTrigger asChild>
-								<span className="inline-flex">
-									<button
-										type="button"
-										className="flex w-8 items-center justify-center rounded-r-lg text-muted-foreground transition-colors hover:bg-muted/80 hover:text-foreground focus-visible:outline-none focus-visible:ring-0 focus-visible:ring-offset-0 disabled:opacity-50"
-										onClick={() => {
-											void handlePickFilesystemRoot();
-										}}
-										disabled={!canAddMoreLocalRoots}
-										aria-label="Add folder"
-									>
-										<FolderPlus className="size-3.5" />
-									</button>
-								</span>
-							</TooltipTrigger>
-							<TooltipContent side="top" className="text-xs">
-								{canAddMoreLocalRoots
-									? "Add folder"
-									: `You can add up to ${MAX_LOCAL_FILESYSTEM_ROOTS} folders`}
-							</TooltipContent>
-						</Tooltip>
-					) : (
-						<button
-							type="button"
-							className="flex w-8 items-center justify-center rounded-r-lg text-muted-foreground transition-colors hover:bg-muted/80 hover:text-foreground focus-visible:outline-none focus-visible:ring-0 focus-visible:ring-offset-0 disabled:opacity-50"
-							onClick={() => {
-								void handlePickFilesystemRoot();
-							}}
-							disabled={!canAddMoreLocalRoots}
-							aria-label="Add folder"
-						>
-							<FolderPlus className="size-3.5" />
-						</button>
-					)}
-				</div>
-			</div>
-			<div className="mx-4 mb-2">
-				<div className="relative flex-1 min-w-0">
-					<div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3 text-muted-foreground">
-						<Search size={13} aria-hidden="true" />
-					</div>
-					<Input
-						ref={localSearchInputRef}
-						className="peer h-8 w-full pl-8 pr-8 text-sm bg-sidebar border-border/60 select-none focus:select-text"
-						value={localSearch}
-						onChange={(e) => setLocalSearch(e.target.value)}
-						placeholder="Search local files"
-						type="text"
-						aria-label="Search local files"
-					/>
-					{Boolean(localSearch) && (
-						<button
-							type="button"
-							className="absolute inset-y-0 right-0 flex h-full w-8 items-center justify-center rounded-r-md text-muted-foreground hover:text-foreground transition-colors"
-							aria-label="Clear local search"
-							onClick={() => {
-								setLocalSearch("");
-								localSearchInputRef.current?.focus();
-							}}
-						>
-							<X size={13} strokeWidth={2} aria-hidden="true" />
-						</button>
-					)}
-				</div>
-			</div>
-			<LocalFilesystemBrowser
-				rootPaths={localRootPaths}
-				searchSpaceId={searchSpaceId}
-				active={currentFilesystemTab === "local"}
-				searchQuery={debouncedLocalSearch.trim() || undefined}
-				onOpenFile={(localFilePath) => {
-					openEditorPanel({
-						kind: "local_file",
-						localFilePath,
-						title: localFilePath.split("/").pop() || localFilePath,
-						searchSpaceId,
-					});
-				}}
-			/>
-		</div>
+		<DesktopLocalTabContent
+			localRootPaths={localRootPaths}
+			canAddMoreLocalRoots={canAddMoreLocalRoots}
+			maxLocalFilesystemRoots={MAX_LOCAL_FILESYSTEM_ROOTS}
+			searchSpaceId={searchSpaceId}
+			onPickFilesystemRoot={handlePickFilesystemRoot}
+			onRemoveFilesystemRoot={handleRemoveFilesystemRoot}
+			onClearFilesystemRoots={handleClearFilesystemRoots}
+			onOpenLocalFile={(localFilePath) => {
+				openEditorPanel({
+					kind: "local_file",
+					localFilePath,
+					title: localFilePath.split("/").pop() || localFilePath,
+					searchSpaceId,
+				});
+			}}
+			electronAvailable={!!electronAPI}
+		/>
 	);
 
 	const documentsContent = (
@@ -1428,7 +1291,7 @@ function AuthenticatedDocumentsSidebar({
 						{cloudContent}
 					</TabsContent>
 					<TabsContent value="local" className="mt-0 flex min-h-0 flex-1 flex-col">
-						{localContent}
+						{currentFilesystemTab === "local" ? localContent : null}
 					</TabsContent>
 				</Tabs>
 			) : (
